修复(worldsync): 进度超过约 21MB 时百分比不再溢出为负数 fix(worldsync): keep progress percent non-negative past about 21MB
- 百分比改为 64 位乘法，避免已下载字节乘 100 溢出成负数
- 同步 Wiki 与进服告示
- 同步服务器运行时状态、物价与市场数据等

- Compute percent with 64-bit multiplication so downloaded bytes times 100 no longer overflow
- Synced the wiki and the join notice
- Synced server runtime state, prices, and market data

+++ tools/mcwws-world-sync/src/main/java/work/mcwws/worldsync/SyncProgress.java
@@ -67,7 +67,7 @@ public final class SyncProgress {
         if (total <= 0) {
             return 0;
         }
-        return Math.min(100, done * 100 / total);
+        return (int) Math.min(100L, done * 100L / total);
     }
 
     private String pctText() {
